+++ src/apis/userInfo/index.ts
@@ -1,0 +1,5 @@
+import instance from 'apis/instance';
+
+export const getMyInfo = () => {
+  return instance.get(`/group`);
+};
